nci-agency/anet#3500: Create separate package for models
