+++ anet-tools/src/core/helper_methods.py
@@ -61,7 +61,7 @@ class helper_methods:
     
     @staticmethod
     def update_anet_models(tables, conn_str):
-        path = "/home/jovyan/work/src/core/models.py"
+        path = "/home/jovyan/work/src/core/objects/core_objects.py"
         generate_command = "sqlacodegen " + conn_str + " --outfile " + path
         if len(tables) != 0:
             generate_command += " --tables " + "'" + ",".join(tables) + "'"
@@ -70,4 +70,4 @@ class helper_methods:
             os.system(generate_command)
             print("Entity classes generated inside " + path)
         except Exception as e:
-            print("EXCEPTION WHILE GENERATING ENTITY CLASSES: ", str(e))
+            print(f"EXCEPTION WHILE GENERATING ENTITY CLASSES: {str(e)}")
